added two files for functions and data used in all pages rating picker, feeling picker is complete, adding same design to all pages
