+++ src/components/App/App.css
@@ -21,3 +21,10 @@
 h4 {
   font-style: italic;
 }
+
+.feelingInputDiv {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
