+++ src/components/FeedbackForm/FormPieces/GetLabel.jsx
@@ -1,0 +1,9 @@
+import labels from "./Labels";
+
+function getLabelText(value) {
+
+    return `${value} Star${value !== 1 ? 's' : ''}, ${labels[value]}`;
+
+}
+
+export default getLabelText;
