+++ src/components/FeedbackForm/FormPieces/Labels.jsx
@@ -1,0 +1,9 @@
+const labels = {
+    1: 'Useless',
+    2: 'Poor',
+    3: 'Ok',
+    4: 'Good',
+    5: 'Excellent',
+  };
+
+  export default labels;
